backend/requirements.txt
fastapi
uvicorn
supabase
google-generativeai
python-multipart
pydantic
pydantic-settings
python-dotenv